refactor(runtime): relocate RichRecord* request types to proximadb-runtime (TD-104 Phase 0)
Move RichRecordBatchRequest / RichRecordDeleteBatchRequest / RichRecordGetRequest
from the root crate (src/services/operations/vectors/legacy.rs) DOWN to
proximadb-runtime (new rich_record.rs), with a root re-export shim so all existing
import paths are unchanged (zero call-site churn). Mirrors the proven S3a
BatchOperationResult relocation pattern (src/services/operations/batch_result.rs).

Why: the RecordOpsPort contract — which proximadb-runtime owns — needs these
types in scope so the port surface is self-contained in the runtime crate (the
terminal ROOT deletion, S3-c, requires RecordOpsPort to not reach up into root).
The only non-primitive field is Vec<ProximaRecord>; runtime already depends on
proximadb-records (added in S3a).

This is TD-104 ROOT-deletion Phase 0 — the cross-crate prerequisite that unblocks
S3-c (Arrow Flight runtime-backed RecordOpsPort). Tracked in the TD-104 ROOT-deletion
roadmap (project memory).

Verified: cargo clippy --lib --bins (-D warnings) + cargo check --tests clean.
